+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for 1.x version
 
+## 2.0.0
+
+- [feature] Made `KeyFactory#doGenerate` protected in order to override only the generation part.
+
+- [minor] The length computation `actual_length / 2` for the random key will be done in the `KeyFactory` instead of the extension.
+
 ## 1.2.0
 
 - [feature] made hashing services configurable: (#33)